MISC: Move orientation_helpers to pose_editor.py
